More pointers, arrays and strings

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -15,7 +15,7 @@ void print_number(int n)
 	}
 	if (n >= 10)
 	{
-		print_number(n / 10);
+		print_number((unsigned) n / 10);
 	}
-	_putchar(n % 10 + '0');
+	_putchar((unsigned) n % 10 + '0');
 }
